more styles added and mobile responsiveness added for other pages

+++ src/component/Item/Item.js
@@ -4,13 +4,17 @@ import { useNavigate } from 'react-router-dom';
 const Item = ({ fruit }) => {
     const { _id, name, img, description, price, quantity, supplier } = fruit;
     const handleRemove = () => {
-        fetch(`https://thawing-hollows-22749.herokuapp.com/fruits/${_id}`, {
-            method: 'DELETE'
-        })
-            .then(res => res.json())
-            .then(data => {
-                alert('Item removed successfully');
+        const agree = window.confirm('Are you sure you want to delete?');
+
+        if (agree) {
+            fetch(`https://thawing-hollows-22749.herokuapp.com/fruits/${_id}`, {
+                method: 'DELETE'
             })
+                .then(res => res.json())
+                .then(data => {
+                    alert('Item removed successfully');
+                })
+        }
     }
 
     const navigate = useNavigate();
@@ -20,8 +24,8 @@ const Item = ({ fruit }) => {
     }
     return (
 
-        <tr className=''>
-            <th scope="row"><img className='rounded-circle' src={img} alt="" width={100} /></th>
+        <tr className='table-row'>
+            <th scope="row"><img className='' src={img} alt="" width={100} /></th>
             <td>{name}</td>
             <td>{supplier}</td>
             <td>{price}</td>
@@ -29,7 +33,7 @@ const Item = ({ fruit }) => {
             <td>
                 <div className='d-flex flex-column'>
                     <button onClick={handleRemove} className='btn btn-danger mb-2'>Remove Item</button>
-                    <button onClick={() => handleUpdate(_id)} className='btn btn-primary'>Update Item</button>
+                    <button onClick={() => handleUpdate(_id)} className='btn btn-update'>Update Item</button>
                 </div>
             </td>
         </tr>
